dqn drqn

+++ environment/gym_env_continuous.py
@@ -60,9 +60,9 @@ class ContinuousPsoGymEnv(gym.Env):
 
     def simple_reward(self, difference):
         if difference > 0:
-            return 1
+            return np.float32(1)
         else:
-            return self._penalty_for_negative_reward
+            return np.float32(self._penalty_for_negative_reward)
 
     def difference_reward(self, difference):
         return max(difference, self._penalty_for_negative_reward)
